requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
asyncpg==0.29.0

redis==5.0.8

pydantic-settings==2.6.1

pytest==9.0.2